main: just realized the size of a signed integer needs to be larger than 8 bits to store more than -128 to +127

+++ Faces/pinball.h
@@ -57,7 +57,7 @@ void WatchyFaceX::drawFacePinball(bool enableDarkMode) {
     }
 
     // new temp variable to prevent calculation underflows
-    int8_t newCoordinate = 0;
+    int16_t newCoordinate = 0;
 
     switch (direction) {
     case DIRECTION_DISP_DOWN:
